requirements.txt
aws-cdk-lib==2.138.0
constructs>=10.0.0,<11.0.0
python-dotenv>=0.15.0,<1.0.0
twilio>=6.26.3,<7.0.0
boto3>=1.18.0,<2.0.0
